Resolved issue #226
