+++ src/interfaces/arguments/Arguments.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Darling\Rig\interfaces\arguments;
+
+/**
+ * Description of this interface.
+ *
+ */
+interface Arguments
+{
+    /** @return array<string, string> */
+    public function asArray(): array;
+
+}
+
